Update gradle to version 6.9
The project was using gradle 4.1 which was released in August 2017. This
version of gradle doesn't cope well with modern JDKs (11 and greater).

The upgrade was done by copying the files from another project to bootstrap
a modern gradle and to update it with :

  cp -a ../code-test/{gradle,gradlew} .
  ./gradlew wrapper --gradle-version 6.9

While today gradle 7.1 is available the project is currently using
deprecated gradle features not compatible with gradle 7.0 and higher. This
is why 6.9 has been selected as gradle version.